pyproject.toml
[tool.poetry]
name = "Google-reviews-analyzer"
version = "0.1.0"
description = ""
authors = ["eliagonzalezmolina <elia.gonzalez.molina@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pre-commit = "^3.6.2"
dvc = "^3.48.4"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"
openai = "^1.14.0"
pandas = "^2.2.1"
tiktoken = "^0.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
